+++ api/src/test/java/org/testifyproject/trait/MethodTraitTest.java
@@ -16,6 +16,7 @@
 package org.testifyproject.trait;
 
 import java.lang.reflect.Method;
+import java.lang.reflect.Parameter;
 import java.lang.reflect.Type;
 import java.util.List;
 import java.util.Optional;
@@ -59,6 +60,14 @@ public class MethodTraitTest {
         List<Class> result = sut.getParameterTypes();
 
         assertThat(result).containsExactly(String.class);
+    }
+
+    @Test
+    public void callToGetParametersShouldReturn() {
+        List<Parameter> result = sut.getParameters();
+
+        assertThat(result).hasSize(1);
+        assertThat(result.get(0).getType()).isEqualTo(String.class);
     }
 
     @Test