+++ junit4/junit4-core/src/test/java/org/testifyproject/junit4/fixture/collaborator/NamedCollaboratorProvider.java
@@ -13,36 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.testifyproject.fixture;
+package org.testifyproject.junit4.fixture.collaborator;
 
 import org.testifyproject.annotation.Name;
+import org.testifyproject.junit4.fixture.common.Hello;
 
 /**
  *
  * @author saden
  */
-public class MethodTestService implements TestContract {
+public class NamedCollaboratorProvider {
 
-    private String message = "Hi!!!";
-
-    void init() {
-        message = "init";
+    String provide(@Name("hello") Hello hello) {
+        return hello.greet();
     }
 
-    @Name("kill")
-    void destroy() {
-        message = "destroy";
-
+    @Name("hello")
+    Hello namedHelloProvider() {
+        return new Hello();
     }
 
-    @Override
-    public String getMessage() {
-        return String.format(message);
-    }
-
-    @Override
-    public String sayHello(String name) {
-        return String.format("Hello %s!", name);
+    Hello unnamedHelloProvider() {
+        return new Hello();
     }
 
 }